Continued WIP on score refactor
